Fixed implicit casts from size_t to int.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@578 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Graphics/Mesh3D.cpp
@@ -108,7 +108,7 @@ int Mesh3D::addPoint(double x, double y, double z, double texX, double texY) {
    if (y < xMin) yMin = y;
    if (z > xMax) zMax = z;
    if (z < xMin) zMin = z;
-   return points.size();
+   return (int) points.size();
 }
 
 /**
@@ -123,7 +123,7 @@ int Mesh3D::addPoint(double x, double y, double z) {
    if (y < xMin) yMin = y;
    if (z > xMax) zMax = z;
    if (z < xMin) zMin = z;
-   return points.size();
+   return (int) points.size();
 }
 
 int Mesh3D::addPoint(MeshPoint* p) {
